Update README.md
Submit your website write-up below here and push back to your branch
there are a few sights that come to mind when I think of mass users and form vs functionality.
examplt I brought up in class was netflix.   Reason why this sight works in my example is the idea that the design is made for" stupid" people
I cannot cound how many times drunk or "non-computer savvy" people complained about the UX just because there was such a breakdown between designers and "commen people"
People need to be able to navigate a sight with one finer and one eye. period.Plans to make a home shopping network UX is in the works.
3 clicks select. agree. ship.  (sold $$$)

on the flip side   reddit.
quick info in a list form interface.
mass info.
adhd
always changing
it works but its not pretty.  but it can be if you find it


in conclution,  a sigt is designed for its users.  the same as a building.
if you build it, they will come.   
just build it for the people you plan to house.